Mo' betta logging, rmdir guard and license
Added slightly more robust logging flexibility throughout the library.
Now the 'binary' accepts verbosity levels from 0 - 2. The default level
is 1, and level 0 disables output while level 2 is probably only useful
for debugging purposes.

Now when using utility.rmdir, an existance check is performed before the
attempt to remove the directory is made. The assumption here is that an
attempt to remove a non-existant directory is not exception worthy.

Bound is not distributed under the MIT license.